requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
python-jose[cryptography]
pillow